Angle snapping works
