+++ src/main/java/frc/robot/subsystems/swerve/DriveConstants.java
@@ -92,7 +92,7 @@ public class DriveConstants {
       };
 
   public static final HeadingControllerConstants HEADING_CONTROLLER_CONSTANTS =
-      new HeadingControllerConstants(2.7, 0, 5, 200, 0.002);
+      new HeadingControllerConstants(3, 0, 5, 200, 0.002);
 
   public static final double[] REEF_SNAP_ANGLES = {-120, -60, 0, 60, 120, 180};
