+++ src/main/java/frc/robot/subsystems/swerve/controllers/HeadingController.java
@@ -8,8 +8,6 @@ import edu.wpi.first.math.trajectory.TrapezoidProfile.Constraints;
 import edu.wpi.first.math.util.Units;
 import frc.robot.Constants;
 import frc.robot.RobotState;
-import frc.robot.subsystems.swerve.DriveConstants;
-
 import org.littletonrobotics.junction.AutoLogOutput;
 
 public class HeadingController {
